requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
httpx>=0.27.0
python-dateutil>=2.9.0
PyYAML>=6.0.2
